Fix exception on parsing empty chunk at end of file
Only occurred when passing an fully empty chunk to polars (i.e. when
the chunk edges align with the full file). For example, when
n_per_chunk=10 and there are 100 events, it would fail on the 11th chunk
that is empty and should be interpreted to end the parsing. In principle,
this caused no problems, as it would only raise it exception when it was
finished. However, it's best to have it complete cleanly.

Also added messages to make it clear that the parsing was successful