Warn more loudly when we reach unexpected behavior
This is entirely fine for polars. If we get it in our cases, it's
unexpected and should be checked by the user. We can safely raise the
exception because we already know the parsing has reached the end of the
file, so exiting due to the exception won't lead to data loss